LayoutTests/accessibility/add-children-pseudo-element-expected.txt

Language Email
Make sure that we are updating the render block flow element's children correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.childrenCount is 3
PASS element.childrenCount is 2
PASS successfullyParsed is true

TEST COMPLETE
